Reenable connect checks for aws and azure
It was disabled in f40a93c90, possibly by mistake because we set a
variable used only when connecting to c8y. However, aside from setting
that variable we emit a warning if the check is not successful, which is
desired. As such, connection checks for these clouds were reenabled.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>