fix: aws and az fail when no response on connection check
When connecting to aws and azure, during "Verifying device is connected
to cloud..." stage, if we don't get response from a remote broker, mark
the spinner as failed instead of OK.

It was being marked as OK, because the code was a bit too c8y-centric:
for c8y we connect directly to the remote broker to create a device and
after that switch to the bridge. When using the bridge, if remote broker
doesn't respond, we know that the bridge is connected with remote broker
because it's using the same configuration as the direct connection,
which was successful, so we consider device status unknown instead of
failing the connection.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>