Fix memory leak
The first design was naive (i.e. loading the javascript module
for each and every function call) and leading to a memory leak
(loading a module with the same name keeps the previous version of the
module in memory).

The new design fixes the issue. The memory is stable while translating to c8y
600 tedge measurements per second during 20 minutes (mode debug)

```shell
$ watch ps -p 1382888 -o args,%cpu,etimes,times,%mem,rss,vsz
COMMAND                     %CPU ELAPSED     TIME %MEM   RSS    VSZ
tedge-mapper gen            47.4    1322      627  0.0 41440 1359476

$ tedge mqtt sub 'test/output' --duration 1000 | pv | wc -l
89.3MiB 0:16:40 [91.4KiB/s] [  <=>                                                                                                                            ]
586311
```

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>