feat: tedge cert renew c8y w/ HSM
This commit introduces support for tedge cert renew c8y when the private
key is located on the HSM.

To generate a CSR, one needs to put the public key in the CSR (it will
become a part of the certificate (SPKI)) and to sign the certificate
using the private key to prove the ownership of it. Because the public
key can be derived from the private key, and because we need to sign
anyway, until this PR we just read the private key to do both these
things.

Now, with an HSM, we can't read the private key, so instead for CSR
creation with HSM being enabled, we read the public key from the
currently used certificate and sign using the TedgeP11Client.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>