fix wrong signature scheme being used for CSR gen
When writing signature algorithm to CSR, we used the signatureAlgorithm
value from the current certificate. That was wrong: signatureAlgorithm
in the certificate describes the signature made by the CA, not our
private key. So when submiting CSR for renewal, CA uses its privkey (C8Y
CA used ECDSA256) and we get a cert where sigAlg is always ECDSA256.

So when we used cert renew on a self-signed cert it always worked
because we CSR sigAlg == cert sigAlg but when renewing a certificate
that was CA signed, we read CA's sigAlg (ECDSA256) but signed with a
different algorithm if we had a different key than ECDSA256

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>