feat: create parent directories if missing for config_update operation
Previously, the `config_update` operation would fail if any parent
directories of the target file were missing.

This change ensures that missing parent directories are automatically
created when the operation is executed.

The mode, user, and group of the immediate parent directory can be
configured via `tedge-configuration-plugin.toml` using the following
keys:
- `parent_mode`
- `parent_user`
- `parent_group`

If `parent_user` and `parent_group` are not explicitly specified,
but the target file's ownership is provided, the parent directory will
inherit the file's user and group.

Signed-off-by: Rina Fujino <rina.fujino.23@gmail.com>