fix: pkcs11 generate correct CSR signature
Make the signer use a PKCS 1.5 signature when signing CSRs instead of
currently used RSA-PSS.

The signature generated so far was invalid because, when generating a
CSR with a private key on a PKCS11 token, and if that key was an RSA
key, the signer signed the CSR using RSA-PSS signature scheme, but the
SignatureAlgorithm identifier in the CSR indicated that the signature
was a PKCS 1.5 signature.

This was previously unnoticed because Cumulocity did not verify CSR
signature is correct when renewing, but as of time of making this fix,
Cumulocity now correctly verifies the signatures.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>