Use a flat list of measurement metadata
Mimicing the structure of grouped measurements as in `"Foo": {"unit: "u", "Temperature": {"unit": "°C"}}`
leads to confusion as units are assigned to `"Foo":20` values as well as `"Foo" : { "Temperature": 20 }`.

Using a flat structure is less confusing as units for `"Foo"` and `"Foo.Temperature"`
are now defined using two independent property sets.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>