pkcs11 add dedicated request type for reloading
TedgeP11Client::with_ready_check used to make a ChooseScheme request
that it immediately discarded only to make sure the library is loaded.
That made unnecessary operation and generated output that was also
unnecessary and made the log harder to read.

Now with_ready_check uses a separate Ping request to which the server
responds by only ensuring the library is loaded and doing nothing else.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>