deps[tedge-p11-server]: Add secrecy-0.8.0
This version of secrecy is already used by cryptoki, so we don't
actually pull anything more, but need to declare it because cryptoki
doesn't reexport it.

Current version of secrecy is 0.10.0, but cryptoki doesn't use it yet,
so using this old version to not to have to pull in 2 versions. Can
upgrade once cryptoki does as well.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>