Reconnect flow input on end-of-stream
`tail -F` can be used to follow the content of a file even when the file
is removed and recreated. However, when the file is a named pipe
(created with mkfifo), `tail -F` crashes with an error:
`cannot seek to offset 0: Illegal seek`, disconnecting the flow from its
input: the process is dead.

=> On end-of-stream, tedge flows respawns the input process.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>