Don't consider the operation failed if saving key_uri fails
Even if for some reason saving URI to tedge-config fails, consider the
operation succeeded if the key was created. Saving URI is only a
convenience and the user can write the URI themselves.

Also extract device.key_uri extraction to separate function, modify
comments to be more helpful by proving more context why it exists and
what should be done with it.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>