deps: Add features to workspace tokio definition
Enumerate used tokio features in the workspace tokio definition, so that
when running tests on member crates that don't specify the features
(tedge-p11-server) the tests don't fail because of missing features.

I figure it makes more sense to put the features in one place instead of
each crate specifying features separately, seeing as after build all the
crates use the same build of tokio anyway.

Signed-off-by: Marcel Guzik <marcel.guzik@cumulocity.com>