KV store shared by all the flows
This is first step toward a context used by flow scripts to access shared data,
such as entity metadata or measurement units. For this first step,
the store is not passed as an argument to the onMessage and onInterval method.
The scripts has to create a `new FlowStore()` providing a `get()` method
to get the value attached to some key - typically the metadata attached to an MQTT topic.

The scripts only have a read access to the shared store
which is populated by regular flows which output is configured to be the context.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>